Update to 3.32.2
**Changelog:**

- Bug fixes and server compatibility improvements
- User interface translation updates

**Note:**

Henceforth, Geary is held back to 3.32 series. Geary 3.34+ requires libhandy for "responsive" design / mobile support, which isn't a target for Solus. In my view (as the Experience Lead and maintainer of our GNOME Stack) this is not acceptable, as we focus solely on providing an optimal modern home computing (read: desktops, laptops, etc.) experience and thus application experiences we provide should cater towards that as well.

Should Geary no longer successfully compile in the future against newer libraries, Vala, etc. it will be deprecated from the repository.